Added buffer overflow for Praetorian Challenge
This is the buffer overflow that worked to get the flag in the Praetorian tech challenge. It uses format string vulnerability to get the start location in memory in order to bypass ASLR in mode 1 or 2. It does a callback to a server in Digital Ocean on Port 4444. Once it does you have a netcat shell.